pyproject.toml
[project]
name = "comic-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.52.0",
    "pillow>=12.0.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
]
